Refactored out Header/Footer into html files + more
Most pages now reference /_templates/header.html or
/_templates?footer.html to load in the header and footer.
So if I DO change the header or footer I can do it in one
place.

Removed style.css and moved stuff to main.css. Not sure if this
was a good idea. The idea being that I should separate MY
changes from the THEME changes. Yeah, I'm gonna roll that change
back.

Fixed the typos mentioned on Facebook.
Fixed the missing link mentioned on Facebook.
Added img-bordered CSS style and p.caption style
to improve the appearance of the portfolio page.